move the ToolServer class into the existing tool_servers folder (#359)
