+++ python/src/kagent/tool_servers/__init__.py
@@ -1,9 +1,11 @@
 from ._ssemcptoolserver import SseMcpToolServer, SseMcpToolServerConfig
 from ._stdiomcptoolserver import StdioMcpToolServer, StdioMcpToolServerConfig
+from ._tool_server import ToolServer
 
 __all__ = [
     "SseMcpToolServer",
     "SseMcpToolServerConfig",
     "StdioMcpToolServer",
     "StdioMcpToolServerConfig",
+    "ToolServer",
 ]
