+++ python/src/kagent/tool_servers/_ssemcptoolserver.py
@@ -3,7 +3,7 @@ from autogen_ext.tools.mcp._config import SseServerParams
 from autogen_ext.tools.mcp._factory import mcp_server_tools
 from loguru import logger
 
-from tool_servers import ToolServer
+from ._tool_server import ToolServer
 
 
 class SseMcpToolServerConfig(SseServerParams):
