--- python/src/tool_servers/__init__.py
@@ -1,3 +1,0 @@
-from ._tool_server import ToolServer
-
-__all__ = ["ToolServer"]
